updated for adc ins on pa1, pc3, and pc4 (using DMA)
16 entry wavetables can now be defined and played back arbitrarily

the adc reading for the "morph" control is converted to 16 bit int for
the interpolation function by bit shifting to the left by 4. 

this is used to interpolate between the two entries in the wavetable
family as specified in the variable declaration section for attack and
release.

to do: 1voct scaling (lookup table?), mode switching